Make subscribing for foreground/background events optional
This lets us use the PasscodeLockViewController in situations other than
the first window that is displayed upon app open. For an example,
consider an app that requires passcode authentication for certain
actions, and which keeps the passcode view controller around. With
events always enabled, the passcode view controller would trigger
biometric login even if it were not being displayed.